[Add] network intial version
