+++ perimeter81/data_source_networks.go
@@ -240,11 +240,7 @@ func dataSourceNetworksRead(ctx context.Context, d *schema.ResourceData, m inter
 
 	networks, _, err := client.NetworksApi.GetNetworks(ctx)
 	if err != nil {
-		diags = append(diags, diag.Diagnostic{
-			Severity: diag.Error,
-			Summary:  "Unable to create Network",
-			Detail:   err.Error(),
-		})
+		return appendErrorDiags(diags, "Unable to get Networks", err)
 	}
 
 	newNetworks := flattenNetworksData(networks.Networks)
@@ -255,98 +251,4 @@ func dataSourceNetworksRead(ctx context.Context, d *schema.ResourceData, m inter
 	d.SetId(strconv.FormatInt(time.Now().Unix(), 10))
 
 	return diags
-}
-
-func flattenNetworksData(networkItems []perimeter81Sdk.Network) []interface{} {
-	if networkItems != nil {
-		networks := make([]interface{}, len(networkItems))
-		for i, serverItem := range networkItems {
-			network := make(map[string]interface{})
-			network["name"] = serverItem.Name
-			network["id"] = serverItem.Id
-			network["tags"] = serverItem.Tags
-			network["subnet"] = serverItem.Subnet
-			network["dns"] = serverItem.Dns
-			network["accesstype"] = serverItem.AccessType
-			network["isdefault"] = serverItem.IsDefault
-			network["tenantid"] = serverItem.TenantId
-			network["createdat"] = serverItem.CreatedAt
-			network["updatedat"] = serverItem.UpdatedAt
-			network["regions"] = flattenNetworkRegionsData(serverItem.Regions)
-			networks[i] = network
-		}
-		return networks
-	}
-	return make([]interface{}, 0)
-}
-
-func flattenNetworkRegionsData(regionItems []perimeter81Sdk.NetworkRegion) []interface{} {
-	if regionItems != nil {
-		regions := make([]interface{}, len(regionItems))
-		for i, regionItem := range regionItems {
-			region := make(map[string]interface{})
-			region["network"] = regionItem.Network
-			region["dns"] = regionItem.Dns
-			region["name"] = regionItem.Name
-			region["tenantid"] = regionItem.TenantId
-			region["createdat"] = regionItem.CreatedAt
-			region["updatedat"] = regionItem.UpdatedAt
-			region["id"] = regionItem.Id
-			region["instances"] = flattenNetworkInstancesData(regionItem.Instances)
-			regions[i] = region
-		}
-		return regions
-	}
-	return make([]interface{}, 0)
-}
-
-func flattenNetworkInstancesData(instanceItems []perimeter81Sdk.NetworkInstance) []interface{} {
-	if instanceItems != nil {
-		instances := make([]interface{}, len(instanceItems))
-		for i, instanceItem := range instanceItems {
-			instance := make(map[string]interface{})
-			instance["network"] = instanceItem.Network
-			instance["dns"] = instanceItem.Dns
-			instance["tenantid"] = instanceItem.TenantId
-			instance["createdat"] = instanceItem.CreatedAt
-			instance["updatedat"] = instanceItem.UpdatedAt
-			instance["resourceid"] = instanceItem.ResourceId
-			instance["ip"] = instanceItem.Ip
-			instance["id"] = instanceItem.Id
-			instance["imageversion"] = instanceItem.ImageVersion
-			instance["imagetype"] = instanceItem.ImageType
-			instance["region"] = instanceItem.Region
-			instance["instancetype"] = instanceItem.InstanceType
-			instance["tunnels"] = flattenNetworkTunnelsData(instanceItem.Tunnels)
-			instances[i] = instance
-		}
-		return instances
-	}
-
-	return make([]interface{}, 0)
-}
-
-func flattenNetworkTunnelsData(tunnelItems []perimeter81Sdk.NetworkTunnel) []interface{} {
-	if tunnelItems != nil {
-		tunnels := make([]interface{}, len(tunnelItems))
-		for i, tunnelItem := range tunnelItems {
-			tunnel := make(map[string]interface{})
-			tunnel["instance"] = tunnelItem.Instance
-			tunnel["interfacename"] = tunnelItem.InterfaceName
-			tunnel["leftallowedip"] = tunnelItem.LeftAllowedIP
-			tunnel["leftendpoint"] = tunnelItem.LeftEndpoint
-			tunnel["network"] = tunnelItem.Network
-			tunnel["region"] = tunnelItem.Region
-			tunnel["requestconfigtoken"] = tunnelItem.RequestConfigToken
-			tunnel["type"] = tunnelItem.Type_
-			tunnel["id"] = tunnelItem.Id
-			tunnel["tenantid"] = tunnelItem.TenantId
-			tunnel["createdat"] = tunnelItem.CreatedAt
-			tunnel["updatedat"] = tunnelItem.UpdatedAt
-			tunnels[i] = tunnel
-		}
-		return tunnels
-	}
-
-	return make([]interface{}, 0)
 }
