+++ perimeter81/resource_network.go
@@ -2,8 +2,6 @@ package perimeter81
 
 import (
 	"context"
-	"fmt"
-	"strings"
 	perimeter81Sdk "terraform-provider-perimeter81/perimeter81sdk"
 	"time"
 
@@ -76,12 +74,14 @@ func resourceNetwork() *schema.Resource {
 func resourceNetworkCreate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 	client := m.(*perimeter81Sdk.APIClient)
+
 	if ctx == nil {
 		ctx = context.Background()
 	}
+
 	network := d.Get("network").([]interface{})[0].(map[string]interface{})
 	name := network["name"].(string)
-	tags := flattenTagsData(network["tags"].([]interface{}))
+	tags := flattenStringsArrayData(network["tags"].([]interface{}))
 	regions := flattenRegionsData(d.Get("region").([]interface{}))
 
 	CreateNetworkPayload := perimeter81Sdk.CreateNetworkPayload{
@@ -95,29 +95,17 @@ func resourceNetworkCreate(ctx context.Context, d *schema.ResourceData, m interf
 
 	status, _, err := client.NetworksApi.NetworksControllerV2NetworkCreate(ctx, DeployNetworkPayload)
 	if err != nil {
-		diags = append(diags, diag.Diagnostic{
-			Severity: diag.Error,
-			Summary:  "Unable to create Network",
-			Detail:   err.Error(),
-		})
+		return appendErrorDiags(diags, "Unable to create Network", err)
 	}
 
-	statusUrl := strings.Split(status.StatusUrl, "/")
-	statusId := statusUrl[len(statusUrl)-1]
+	statusId := getIdFromUrl(status.StatusUrl)
 
 	var networkId string
 	for {
-		networkStatus, _, err := client.NetworksApi.NetworksControllerV2Status(ctx, statusId)
-		if err != nil {
-			diags = append(diags, diag.Diagnostic{
-				Severity: diag.Error,
-				Summary:  "Unable to get Network Status",
-				Detail:   err.Error(),
-			})
-		}
+		var networkStatus perimeter81Sdk.AsyncOperationStatus
+		networkStatus, diags, _ = checkNetworkStatus(ctx, statusId, *client, diags)
 		if networkStatus.Completed {
-			networkUrl := strings.Split(networkStatus.Result.Resource, "/")
-			networkId = networkUrl[len(networkUrl)-1]
+			networkId = getIdFromUrl(networkStatus.Result.Resource)
 			break
 		}
 		time.Sleep(60 * time.Second)
@@ -127,28 +115,31 @@ func resourceNetworkCreate(ctx context.Context, d *schema.ResourceData, m interf
 
 	return diags
 }
+
 func resourceNetworkRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 	client := m.(*perimeter81Sdk.APIClient)
+
 	if ctx == nil {
 		ctx = context.Background()
 	}
+
 	networkId := d.Id()
-	network, _, err := client.NetworksApi.NetworksControllerV2NetworkFind(ctx, networkId)
+	networkData, _, err := client.NetworksApi.NetworksControllerV2NetworkFind(ctx, networkId)
 
 	if err != nil {
-		return diag.FromErr(err)
+		return appendErrorDiags(diags, "Unable to find Network", err)
 	}
 
 	CreateNetworkPayload := perimeter81Sdk.CreateNetworkPayload{
-		Name:   network.Name,
-		Tags:   network.Tags,
-		Subnet: network.Subnet,
+		Name:   networkData.Name,
+		Tags:   networkData.Tags,
+		Subnet: networkData.Subnet,
 	}
 
-	n := flattenNetworkData([]perimeter81Sdk.CreateNetworkPayload{CreateNetworkPayload})
-	if err := d.Set("network", n); err != nil {
-		return diag.FromErr(err)
+	network := flattenNetworkData([]perimeter81Sdk.CreateNetworkPayload{CreateNetworkPayload})
+	if err := d.Set("network", network); err != nil {
+		return appendErrorDiags(diags, "Unable to set Network data", err)
 	}
 
 	return diags
@@ -156,6 +147,7 @@ func resourceNetworkRead(ctx context.Context, d *schema.ResourceData, m interfac
 func resourceNetworkUpdate(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 	client := m.(*perimeter81Sdk.APIClient)
+
 	if ctx == nil {
 		ctx = context.Background()
 	}
@@ -164,22 +156,21 @@ func resourceNetworkUpdate(ctx context.Context, d *schema.ResourceData, m interf
 		networkId := d.Id()
 		network := d.Get("network").([]interface{})[0].(map[string]interface{})
 		name := network["name"].(string)
-		tags := flattenTagsData(network["tags"].([]interface{}))
+		tags := flattenStringsArrayData(network["tags"].([]interface{}))
+
 		networkDto := perimeter81Sdk.BaseNetworkDto{Name: name, Tags: tags}
 		updateNetworkDto := perimeter81Sdk.UpdateNetworkDto{Network: &networkDto}
 		_, _, err := client.NetworksApi.NetworksControllerV2NetworkUpdate(ctx, updateNetworkDto, networkId)
 		if err != nil {
-			diags = append(diags, diag.Diagnostic{
-				Severity: diag.Error,
-				Summary:  "Unable to update Network",
-				Detail:   err.Error(),
-			})
+			return appendErrorDiags(diags, "Unable to update Network", err)
 		}
+
 		d.Set("last_updated", time.Now().Format(time.RFC850))
 	}
 
 	return resourceNetworkRead(ctx, d, m)
 }
+
 func resourceNetworkDelete(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
 	client := m.(*perimeter81Sdk.APIClient)
@@ -188,57 +179,9 @@ func resourceNetworkDelete(ctx context.Context, d *schema.ResourceData, m interf
 	}
 	networkId := d.Id()
 	_, _, err := client.NetworksApi.NetworksControllerV2NetworkDelete(ctx, networkId)
-
 	if err != nil {
-		return diag.FromErr(err)
+		return appendErrorDiags(diags, "Unable to delete network", err)
 	}
 	d.SetId("")
-
 	return diags
-}
-
-func flattenTagsData(tags []interface{}) []string {
-	tagsData := make([]string, len(tags))
-	for i, v := range tags {
-		tagsData[i] = fmt.Sprint(v)
-	}
-	return tagsData
-}
-
-func flattenRegionsData(regionItems []interface{}) []perimeter81Sdk.CreateRegionInNetworkload {
-	if regionItems != nil {
-		regions := make([]perimeter81Sdk.CreateRegionInNetworkload, len(regionItems))
-
-		for i, regionItem := range regionItems {
-			region := perimeter81Sdk.CreateRegionInNetworkload{}
-
-			region.CpRegionId = regionItem.(map[string]interface{})["cpregionid"].(string)
-			region.InstanceCount = int32(regionItem.(map[string]interface{})["instancecount"].(int))
-			region.Idle = regionItem.(map[string]interface{})["idle"].(bool)
-			regions[i] = region
-		}
-
-		return regions
-	}
-
-	return make([]perimeter81Sdk.CreateRegionInNetworkload, 0)
-}
-
-func flattenNetworkData(networkItems []perimeter81Sdk.CreateNetworkPayload) []interface{} {
-	if networkItems != nil {
-		networks := make([]interface{}, len(networkItems))
-
-		for i, networkItem := range networkItems {
-			network := make(map[string]interface{})
-
-			network["name"] = networkItem.Name
-			network["tags"] = networkItem.Tags
-			network["subnet"] = networkItem.Subnet
-			networks[i] = network
-		}
-
-		return networks
-	}
-
-	return make([]interface{}, 0)
 }
